requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
cloudscraper>=1.2.71
pymongo>=4.6.0
python-dotenv>=1.0.0
